requirements.txt
futu-api>=6.0.0
yfinance>=0.2.0
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
